# src/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Base API",
    "description": "BASE API é um conjunto de códigos em arquitetura NodeJs com Typescript, Express e MongoDb, que são utilizadas como base de novos projetos BackEnd.",
    "contact": {
      "name": "Luis Eduardo Rosa",
      "email": "luis.rosa@ideros.com.br"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3140/"
    }
  ],
...
  "paths": {
...
    "/auth/refresh": {
      "get": {
        "tags": [
          "Autenticação"
        ],
        "summary": "Revalidar token de acesso",
        "description": "Método responsável por receber um token válido e gerar novo",
        "operationId": "refreshToken",
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string"
                    },
                    "validAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "unsuccessful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReturnApi"
                }
              }
            }
          },
          "401": {
            "description": "unauthorized access",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReturnApi"
                }
              }
            }
          },
          "500": {
            "description": "bad request",
            "content": {}
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
    "/auth/{authId}": {
      "post": {
        "tags": [
          "Usuários"
        ],
        "summary": "Atualizar registro por Id",
        "description": "Método responsável por atualizar os dados de um registro existente, buscando por Id. Este método não permtirá o envio dos atributos 'secretKey', 'secretIv' ou 'token'",
        "operationId": "updateAuth",
        "parameters": [
          {
            "name": "authId",
            "in": "path",
            "description": "Id da entidade Auth",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Atributos que podem ser editados - enviar apenas os atributos que serão alterados",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "clientId": {
                    "type": "string"
                  },
                  "authType": {
                    "type": "string"
                  },
                  "changePasswordOnNextLogin": {
                    "type": "boolean"
                  }
                }
              }
            }
          },
          "required": false
        },
        "responses": {
          "201": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Auth"
                }
              }
            }
          },
          "400": {
            "description": "unsuccessful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReturnApi"
                }
              }
            }
          },
          "401": {
            "description": "unauthorized access",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReturnApi"
                }
              }
            }
          },
          "500": {
            "description": "bad request",
            "content": {}
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "x-codegen-request-body-name": "body"
      },
      "delete": {
        "tags": [
          "Usuários"
        ],
        "summary": "Deletar registro por Id",
        "description": "Método responsável por excluir os dados de um registro existente, buscando por Id.",
        "operationId": "deleteAuth",
        "parameters": [
          {
            "name": "authId",
            "in": "path",
            "description": "Id da entidade Auth",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReturnApi"
                }
              }
            }
          },
          "400": {
            "description": "unsuccessful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReturnApi"
                }
              }
            }
          },
          "401": {
            "description": "unauthorized access",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReturnApi"
                }
              }
            }
          },
          "500": {
            "description": "bad request",
            "content": {}
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "x-codegen-request-body-name": "body"
      }
    },
    "/auth/changePassword/{authId}": {
      "post": {
        "tags": [
          "Usuários"
        ],
        "summary": "Alterar senha de acesso",
        "description": "Método responsável por realizar a alteração de senha de acesso, obrigatório o envio da senha antiga e da nova senha.",
        "operationId": "changePassword",
        "parameters": [
          {
            "name": "authId",
            "in": "path",
            "description": "Id da entidade Auth",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Atributos obrigatórios",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "currentPassword": {
                    "type": "string"
                  },
                  "newPassword": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReturnApi"
                }
              }
            }
          },
          "400": {
            "description": "unsuccessful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReturnApi"
                }
              }
            }
          },
          "401": {
            "description": "unauthorized access",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReturnApi"
                }
              }
            }
          },
          "500": {
            "description": "bad request",
            "content": {}
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "x-codegen-request-body-name": "body"
      }
    },
...
  },
  "components": {
    "schemas": {
      "ReturnApi": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "Auth": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "clientId": {
            "type": "string"
          },
          "secretKey": {
            "type": "string"
          },
          "authType": {
            "type": "string",
            "enum": [
              "API",
              "USER",
              "ACCOUNT"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "__v": {
            "type": "integer"
          }
        }
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "No campo 'Value' colar o token recebido no método de 'autenticação', e então clique no botão 'Authorize'",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}